refactor: move codex fallback context logic into codex module
